Jung Park and Jack Release 1 commit. Refactored to .class folder and updated pointers.
