+++ classes/person.py
@@ -1,0 +1,7 @@
+class Person:
+    def __init__(self, name, age, role, password):
+        self.name = name
+        self.age = age
+        self.role = role
+        self.password = password
+
