+++ classes/student.py
@@ -1,0 +1,7 @@
+from classes.person import Person
+
+class Student(Person):
+
+    def __init__(self, name, age, role, password, school_id):
+        Person.__init__(self, name, age, role, password)
+        self.school_id = school_id
